Feature #9990
Improve a little some of the JDBC Connector application code.
Restore the MyDB application and rewrite its code from the JDBC Connector's
one: this first step is about only to read the content of a given table as
well as to apply some filter on its content (like it is does in JDBC
Connector).