Feature #9990: add the functionality to modify a table's row.
For doing, we don't use anymore the programmatic build of an XML form in order
to both render the row modification form and to take into account the SQL types
in the rendering of the HTML inputs by converting them to the predefined types
of the XML form fields.
Instead we use a classic HTML form directly built from the row to be updated
and by taking into account some of the SQL types of each row's field. The form
is rendered directly within a javascript popin, providing a better user
experience (read modern user experience). Because the publishers can update the
content of a database table's row, because the publishers work directly with
a database, the characteristics and the constrains of SQL is no more (and
badly) hidden by the GUI. It is expected the publishers know SQL to update the
rows of a table; we provide only a GUI-way to do that (not a simple SQL
statement editor).