Bug #9194
Update some Silverpeas components (aka applications) with some changes in
Silverpeas-Core: the main change is to structure the objects concerned by a
right access profile into an object: ProfiledObjectId. This object is an
identifier to a given resource in Silverpeas and that is convered directly by a
right profile. AS the referred object can be any type, its type is indicated
by the enum ProfiledObjectType and it's embodied in the ProfileObjectId.