Bug #9194
Up to now, subscriptions in Kmelia was managed as they were only done on a
folder. In this behaviour, the subscription in a whole Kmelia application
instance was done like it is a subscription on the root folder of the Kmelia
instance. This brokes the coherence with others Silverpeas applications in
which the subscription on an application instance is different than a
subscription on a resource managed by the application instance.
So, to keep the same behaviour in the subscriptions in Kmelia than in others
applications, the way the subscriptions are managed in Kmelia is changed:
the subscription on the whole Kmelia instance (that is considered in Kmelia as
a subscription on the root folder) is performed as such in the subscription
service like any others applications and thus the subscriptions are fetched
correctly and coherently whatever the applications.