Bug #10884
The checking a publication is an alias is now implemented into a single place
(in KmeliaPublication)

Check when copying/pasting or moving/pasting a folder or a publication, the
publication is an alias or a true publication. In the former, an alias is
created into the new location (destination of the pasting).