Bug #10884
The way the shortcut to a publication is working is now more coherent.

Previously, in Kmelia, if the alias (aka shortuct in the GUI) of a publication
was set in a topic in the same Kmelia instance than the publication, then the
information of being a shortcut was lost and we never knew if a given
publication in a topic was an original one or an alias, driving to some
unexpected behaviour with path-based operations (copy, move, deletion, ...).

Now, whatever the Kmelia instance, any locations of a publication other than
the original one (the one in which the publication was created or moved or
copied) are considered as a shortcut to the original publication.

For doing, a KmeliaPublication represents hence a publication located in a
given topic or, if no topic is provided, in its original location. By knowing
to which topic a Kmelia publication is related, it is more easy to know if that
publication is an alias. And then, it is more easy to take in charge the
aliases in the path-based operations.