Bug #10884
In Kmelia, when a folder is deleted with its content, the number of items in
the folder parent and in the bin where computed from the previous number of
items. In the case the deleted folder contained at least one shortcut, the
result of the computation is false because a shortcut cannot be removed (hence
put into the bin) but only deleted!
The fix here is to ask to the folder web service the nodes that are implied in
the deletion (the parents of the deleted folder (with the root one) and the
bin) and from the sent back information sets the correct number of items in
each of those nodes. For doing, the FolderResource#getNode(String, String)
method, that is invoked at the web service requesting, calls a new method
in KmeliaService: getFolder(NodePK, String) that returns the details of the
specified node with the number of items it contains and with all of its
children (that are visible by the requester (the user)).

Take into account the change in PublicationService and the new Location
object as being a node carrying the information of being or not an alias.