Bug #9194
Remove the explicit call of the unsubscription when a suggestion box or when a
Kmelia is deleted.

Fix a bug when updating the content of a publication in Kmelia: the notification
mechanism updated the current publication detail in the user session instead of
copying it to modify it to set peculiar attributes for the notification
subsystem (like whether the publication is an alias or not for the folder
accessible by a user or a group of users).