Bug #10993
Fix the bug. For doing, carries the column targeted by a foreign key up to the
renderer of the SQL table referred by that foreign key. Indeed, in some
circumstances, the foreign key can refer other that the primary key of the
referred table and in such a situation the column with the foreign key needs to
be valued with the column value of the selected row targeted by the foreign key
instead of the primary key.