Bug #10905
Update Jenkinsfile so that it builds upon the last build version of Silverpeas
Core when its dependency on it isn't specifically defined through the branch
name or the PR title.
Indeed, in some circumstances, Silverpeas Core in its SNAPSHOT version can to
be not built in the local maven repository of our CI service.