Bug #8896
The name of the subscription activation parameter is now a constant defined in
the ResourceSubscriptionService. Any components that supports such a parameter
have to use it instead of its own parameter name. With such a constant, the
management of subscriptions related to this parameter can be done automatically
by the subscription service.