Bug #11109
Fix bug #11109. In MyDB, several foreign keys on the same table can be now set
when updating or inserting a new row.

Fix another discovered bug: simple default value of fields in MyDB weren't
taken into account.

Add a new feature: computed default value of fields are now taken in charge:
MyDB asks for the database to compute it. As some default value computation
depends on the context on which they are invoked, the default value is now
figuring out on the demand.