Feature #11152
Now Silverpeas is based upon JEE 8.

Specify the Docker container to use to execute the build with Java 11 and
Wildfly 20.

Add a dependency in the MANIFEST.MF for integration tests to take into account
some private package (com.sun, sun) are not accessible by default in Java 11.

In Jenkinsfile, compute the version of the parent pom when a snapshot version
of it is declared in the project's pom.