Feature #11841
NodeDetail implements now Identifiable. So, the getId() method is refined to
satisfy the Identifiable interface et the previous getId() method is renamed to
getLocalId().
BeanTranslation identifier property is now a String and not more an Integer in
order to be more coherent with the business representation of an identifier in
Silverpeas.

NodeDetail implements now I18nContribution. In CMIS, we can use its
ContributionIdentifier as the identifier of its representative CMIS object.

The old Content Management API has been lightly refactored by renaming some of
its classes to a an unambiguous name according to their responsability and the
way they are used.

Kmelia implements the new interface CmisContributionsProvider to expose the user
contributions that have to be exposed in the CMIS objects tree. This provider is
used by the Silverpeas CMIS implementation to know what are the user
contributions managed by an application that have to be exposed in the CMIS
objects tree. Both the topics and the publications of the Kmelia instances are
exposed (the peculiar topics "To Validate", "Unclassified", and "Bin" are
excluded). The root folder isn't exposed as expected by our CMIS implementation.