Improve IT
With JPA 2.2, it is not more required to have custom converters to take
into account the new Java Time API when persisting entities. In order to
be compliant in the way JPA 2.2 handle the Java Time API and the way our
code use this API, some of our entities have been refactored: the
previous OffsetDateTime fields are now Instant fields. By the way, the
code based upon those entities are also updated to take into account the
change.