Feature #11841
Add support in Kmelia for the creation of publications through CMIS.

Replace default SimpleAttachment construction by the use of its builder.
Add support to create a publication in Kmelia through CMIS.
When filtering the node to not return, checks the folder is not for non
visible contributions.

In CmisKmeliaContributionsProvider, filter the publications according to
their visibility.

Add Transactional support in CmisKmeliaContributionsProvider for the
contribution creation methods.