Bug #12479
Now, when a Kmelia instance is duplicated, both its default
classification (for the instance itself as well as for each of
the nodes) and its PdC utilization are also duplicated.

Update copyApplicationDialog.jsp to set the property of copying the
classification onto the PdC of the publications. So, all specific
classification positions of each publication will be duplicated. This
property is always set at true but it is taken into account only when
the publications are duplicated, that is to say if the property of
copying the publication headers is set at true.