Feature #12035: wiring the modification popins on massive publication update.
Fixing the fact that a user can modify a contribution for which he has no right!

Fixing the selected checkboxes which was hidden on publication list.

Fixing into KmeliaSessionController the visibility filter performing which was executed before the list of publication was loaded.

Fixing the modification of always visible publicaton into draft status which was not well handled. The valid publication was updated instead of the clone ones.