Feature #12584
Take into account the module silverpeas-core-web is splitted in two:
- silverpeas-core-rs
- and silverpeas-core-web

Take into account in the Jenkinsfile build descriptor the property
core.version in the pom.xml is valued with a string and no more through
a reference to the project version. Indeed, when some Silverpeas
components are a dependency of an external project with its own
versionning, the project.version property is then the one of the
external project and not the one of Silverpeas-Components: it is a
dynamic property.