Feature #12713
For publications in Kmelia:
because the draftOutDate is always set at null and because the automatic
drafting out process (batch process) isn't more used (anyway it cannot
work because the draftOutDate is always null) since 2011, the process
and the draftOutDate attribute and its handling in the code are all
removed of the code.

Add the basket use in Kmelia.