Feature #12713
KmeliaPublication is now an I18nContribution and supports both
WithPermaLink and WithTumbnail features.

Fix in javascript a bug when removing the unselected publications from
the selection basket.

Delete unused code in KmeliaSessionController

Add support of ApplicationService to Toolbox app. Now the publications
of a toolbox app can be put in the selection basket.