Feature #12713
Adding the possibility to send to itself the template.
Using the transversal mail sending management.

Adding the sending of a letter to itself.

Making that a publisher can not manage a template.

Adding a parameter to set the N last newsletter for users on application instance homepage.

Removing the action 'Registering as Template' which is now obsolete.

Refactoring the newletter portlets...

Adding the possibility to select as fast as possible elements for basket selection container (GED, News and Calendar events).