Feature #13197: implementing new community component.
Inception of a new component project: Community.
The goal of the component is to manage a community of users for a given
resource in Silverpeas, for instance for a collaborative space. The
application provides a way for users to have a look of a space (with
its presentation but not its content) and to give them the opportunity
to join the community of users of this space; that is to say to become
a member of this community.

Fix the name of the directory of the db migration scripts for PostgreSQL