Feature #13197: implementing new community component.
Fix subscription messages in main.jsp

Add for each Silverpeas component a role mapping in their descriptor.

Refine the business model and the REST-based web services.
A community of users is always managed by a given Community application
instance. Such application instance manages only one single community of
users for a given resource (for instance only for a collaborative
space). A community of users defines the set of members of a given
space.