Feature #13197: implementing new community component.
Modifying component descriptor in order to specify the profile mapping between space and component.

Adding an implementation of ComponentInstanceAccessControlExtension dedicated to Community component in order to implement specific rules with respect to the space manager status.

Adding CommunityWebManager in order to centralize functional code between WEB Services and Web Component controller.

Implementing the edition of the Wysiwyg content of the space community.