Feature #13197: implementing new community component.
A collaborative space with a community of users is said to be a
community space.

Add unit and integration tests on the management of a community for a
space.

Add a new business entity: CommunityMember to represent a member of a
community.

Add package-info.java
Add unit and integration tests

Add a status of membership for members of a community of users. The
status indicates whether the membership of a user is waiting for
validation, or whether his membership has been committed, or whether his
membership has been removed.

Add two functions for members:
- one to add a user as a member of a community of users,
- one to remove the membership of a user from a community of users.
In latter, a member is never deleted in the table of members of a
community of users. Only his membership status is updated to
MembershipStatus#REMOVED.

Start to add unit test.

CommunityMember provides a method to get the (highest) role a member
plays in a community space.
Add methods to CommunityOfUsers for:
- add a user as a member of the community,
- remove a member from the community.

Add unit and integration tests to check these new methods.

Fix a bug with the CommunityMember#userId. In database, the userId is an
integer whereas in the entity CommunityMember is has been typed as
String: so now userId is an int.

Add the function to get the members of a community of users with
pagination (members only, not those whose membership is awaiting the
validation of the administrator, nor those removed from the community)

Add the function to get the history of the memberships of a community
with pagination.

Rename CommunityMember to CommunityMembership.
Rename CommunityMemberProvider to CommunityMembershipsProvider
Rename CommunityMemberRepository to CommunityMembershipRepository.
Introduce in CommunityMemberhipRepository the CommunityMembershipsTable
object to filter the access to memberships by the community of users.

Fix a bug with CommunityMembershipsTable#getByUser: only the members
whose membership is committed should be got.

Add CommunityMembershipsProvider#getAll to get the history of
memberships of a community. The history includes all members whatever
their status pending member, actual member, no more member).

Add CommunityMemberProvider#getPending to get the
pending members of a community. This isn't yet tested.

Refine the name of some methods.

Implement the synchronization of the table of members of a community
with the users playing a non-inherited role in the community space.
Indeed, a user is member of a community if he plays a non-inherited role
in the community space. But, as users can be added explicitly in the
backoffice to some roles of such a space, this requires to ensure the
table of members is kept up-to-date. For doing, for instance, each time
a call to a list of members or to the history of membership of the
community is invoked, a synchronization is performed. To avoid
interleaving synchronization for the same community, and to avoid to
overload Silverpeas by a flow of synchronization of a given community, a
timer of synchronization is maintained for each community and this timer
is synchronized between several threads.

Add tests on the pending member adding.
Synchronization takes into account also this: if a user is a pending
member but an admin has added him in a role of the community space then
the membership of the user is committed automatically.
Fix some bugs.
The CommunityMembershipsProvider#get(User) returns now  also the
pending membership.

Add a method to CommunityOfUsers to refuse a pending membership of a
user to the community of users.

Add unit and integration tests.

Rename in CommunityOfUsers:
- getUserRoleOn to getUserRoles
- removeAsMember to removeMembership (to be coherent with the new method
refuseMembership)