Feature #13197: implementing new community component.
Add a new REST web resource: CommunityMembershipResource dedicated to
provide access to the memberships of a given community of users.
Add by the same way the web entities that are transmitted between
resource and client.

Update CommunityWebManager to be used by the REST web resources.

Refine the REST-based web resources to use the CommunityWebManager for
getting the requested business objects.

Update the War builder required for the integration tests, so they can
work without any dependencies issues.

Add integration tests on the web resource CommunityMembershipResource.
Fix some bugs discovered with the integration tests.

Add integration tests on the getting of memberships from the
corresponding REST web resource.

When a Community application is instantiated, it is set automatically as
the home page of the parent space, the space for which it has to manage
a community of users.

When the account of a user is deleted in Silverpeas, then all of his
memberships to the communities in Silverpeas are removed.

Add properties to a community of users:
- the home page (and the type of the home page) for the members of the
community of users.
- the URL of a charter (can be a guide) users have to accept to join the
community of users.

Add unit and integration tests.

Add the update of a community of users through the corresponding REST
web resource.
Add and refine integration tests.

Fix the creation of the indexes for the component Community

When a Community app is instantiated in a given space, automatically the
inheritance of the access rights from its parent spaces is disabled.

The duration between two synchronizations of the user profiles with the
memberships table of the community space can be now parameterized with
the property community.memberships.synchronization in the settings bundle
communitySettings.properties