Feature #13179
In order for an anonymous user, once connected, to be redirected to the
contribution view page from which he has asked to sign in, a
ComponentInstanceRoutingMap has been implemented in some of Silverpeas
components supporting the commenting on the contributions they manage.

Take into account the anonymous session is no more a single session for
all anonymous users but for each of them.