Bug #13388
Fix bug.
The issue was comming from the check of specific rights wasn't good.
To cover all the cases, it is required once the topic is created to
check it has only local rights and hence not inherited ones from a
parent topic. Then, we have to recursively to fetch only the admin
right profile from a parent (the not inherited one or the one of the
component instance itself).