Feature #13464
Rename the following methods of ProcessManagerSessionController:
* getAssignForm to getAssignFormOfActiveStates
* getAssignRecord to getAssignRecordOfActiveStates
* reAssign to reassignInActiveStates
because these methods work on only the active states of a process
instance.

Rename the method of ProcessManagerSessionController:
substituteDefinitely to replaceInAllAssignedStates because it works on
all the states assigned to the user for which all the roles he plays has
to be reassigned.
Add the method reassignInAssignedStates in which the reassignation is
done for all the assigned states of a given process instance.

Refactor the method reassignInActiveStates so that the true operation of
reassignation of the roles of some user to others ones in a set of
states is provided by a private method that can be also used by the new
reassignInAssignedStates method.