Feature #13961
Some codes in Silverpeas Core are now provided by Silverpeas Kernel and
latter is a dependency of Silverpeas Core API.

Take into account of the change.
The codes are:
- annotations
- exceptions
- utility classes
- cache API
- logging API
- bundle API

Take into account also the change in unit test with the
new EnableSilverEnvTest JUnit 5 extension.

Take into account of the remove of the ServiceProvider#getSingleton
methods as the caching of singletons is now automatically performed by
ManagedBeanProvider. This avoid any error of caching not singleton
beans (when the scope of their class has changed).