Feature #14033
Now the creation and the deletion of a community of users is notified.
This can be used for external components to access the newly created
commiunity of users in order to perform some post-processing with the
community.

New CommunityOfUsers#getGroupOfMembers() method. This method returns the
group of members in the community. If such a group doesn't yet exist,
then it is created for the community of users. It is a way to get all
the actual members of a community without having to filter the
memberships by their status.

New CommunityOfUsers#getAll() static method to get all the existing
community of users.