Bug #14471
A publication is a pueculiar contribution in Silverpeas as iti can be
located in several locations in the same or in different component
instances. In this case, we have the original publication along with
its all aliases. This capability is used in the Kmelia application.
As consequence, the publication is only referred uniquely by its
local identifier (its local identifier is its global one), whereas
the others types of contributions are identified uniquely by their
global identifier that is made up of their local identifier and of
the identifier of the application instance they belong to.

Despite this peculiarity, the authorization mechanism in Kmelia takes
the publication by their local identifier and, in a such context, any
access right validation can be incorrect on publication aliases
located in a topic for which specific access rights have been set.

Now, the publications, whatever they are (original one or aliases) are
taken in charge by both their local identifier and the identifier of the
component instance they belong to.