Feature #10542
When a folder and its content (both subfolders and publications) are put
into the bin, the actual operation operated by the KmeliaService is now
a move instead of a simple node detail update.
Rename the confirmation text when emptying the trash can.
When the trash can is emptied, the folders deletion is now taken into
account by the treeview.
Fix a bug with the number of publications displayed with the folder name
computed when a folder is deleted with its whole content.