Bug #14661
Fix the bug. The d&d handling was based upon the .jstree-draggable
class. But since the refactoring within the fix of the bug 14478,
the <span> with this CSS class didn't receive anymore the mouse
event but its <a> parent because this one had a class attribute set.
Without the class attribute in the <a> parent, the <span> element
succeed to receive any mouse click down event on it. Nevertheless, in
order the d&d mechanism works on publications, it was needed to move the
id attribute from the <a> element to the <span class=".jstree-draggable>
element because this publication identifier is fetched from this id on
the dragged HTML node. So, the click event handler of the <a> node
required to be refined to take into account of this (because the
id attribute was also fetched to figure out the identifier of the
publication to redirect the user to.