Bug #14595
Adds specific lifecycle CDI stereotype to some beans in order to be
able, in the futur, to enable the management of the beans by CDI for
only those being annotated with CDI stereotype annotations.

Provide another solution to the fix of the bug based upon the system
event notification mechanism. In ResourcesManager component:
- Add listeners of events related to the remove (directly or not) of
  users from the Manager profile instance (role name 'responsable').
- Add listeners of events related to the remove of a user from a group
  that is declared in the Manager profile (role name 'responsable').
- To allow to update the validators list of the resources managed by the
  component according to the events received by the listeners, the
  ResourceRepository and the ResourceService classes have been updated.
- Add integration tests to ensure the listeners work correctly.