Bug #14732
The regression was introduced with the fix of the bug #14661.
Ensure now the access to a given publication is done correctly by
clicking on its name wrapped by a span tag with at least the class
'pub-link' (a shortcut/alias cannot be d&d like a normal publication).

In the seeAlso.tag, encode for HTML both the description and the title
of a referenced publication.