Feature #10786
Add the KmeliaBinsScheduledPurger to purge periodically the all the bins
of the different Kmelia instances of the items (both publications and
topics) that have been removed (place in the bin) at a date older than a
given delay in days. The recurrence of the task and the delay are set in
the org/silverpeas/kmelia/settings/kmeliaSettings.properties
configuration file.