Feature #10786
In Kmelia, take into account topics and publications can be moved or
copied from/to bins in the same or in another Kmelia instances; their
removal status have to be modified consequently of the move or copy
operation.

In the Kmelia bin purger, take into account of the removal properties of
both NodeDetails and PublicationDetails. For all publications and nodes
previously removed (put in the bin) before this feature, the purger will
erase them automatically (we don't know their removal date so, as such,
we make the assumption their deletion are older than the purge delay
setting).
Fix a bug with the Kmelia instance ids: globalize them to be taken by
NodeService and PublicationService to identify the nodes and the
publications in the different Kmelia instances.