Feature #14962
Implements the CommunityFactory to constructs an instance of the
Community application to manage the members of a community space.
The name and the description of a new community instance are set
automatically by using the l18n bundle of the app.

Refactor CommunityOfUsers to use the new CommunityMembershipService
service.

The Community component isn't anymore visible among others Silvepreas
components. Indeed, this component is to be directly instanciated by
Silverpeas Core when a community space is created.