Bug #14998
When browsing the publications to link, avoid to update in the session
the list of publications in the selected folder.
Now, when the browing of DEMs is done within the context of linking
publications, the list of publications of the selected folder is asked
explicitly by using the KmeliaService instead of using the
KmeliaSessionController which works with the user session.